+++ tests/test_download.py
@@ -23,25 +23,15 @@ def test_maja_search():
 def test_download(output_dir):
     zip_dir = (output_dir / "download" / "zip").rmtree_p().makedirs_p()
     unzip_dir = (output_dir / "download" / "unzip").rmtree_p().makedirs_p()
-    # test for search with empty results
-    tile = "T31TGM"
-    start_date = "2024-09-23"
-    end_date = "2024-09-24"
-    zip_files, unzip_files = maja_download(
-        tile=tile,
-        start_date=start_date,
-        end_date=end_date,
-        zip_dir=zip_dir,
-        unzip_dir=unzip_dir,
-        lim_perc_cloud=50,
-        level="LEVEL2A",
-        dry_run=False)
-
 
     # 31TGM 2018-08-11 is duplicate with cloud_cover [41,52]
-    tile = "T31TGM"
-    start_date = "2018-08-11"
-    end_date = "2018-08-12"
+    # 31TGK 2020-05-22 is duplicate with cloud_cover [11,30]
+    tile = "T31TGK"
+    start_date = "2020-05-22"
+    end_date = "2020-05-23"
+    bands=["B2", "B3", "CLMR2", "CLMR1"]
+    cloud_min = 20
+    cloud_max = 40
 
     # # 31TGM 2017-06-19 is duplicate with cloud_cover [3, 4]
     # tile = "T31TGM"
@@ -54,17 +44,18 @@ def test_download(output_dir):
     # end_date="2023-01-13",
     
     # should download duplicates
-    zip_files, unzip_files = maja_download(
+    downloaded, unzip_files = maja_download(
         tile=tile,
         start_date=start_date,
         end_date=end_date,
         zip_dir=zip_dir,
         unzip_dir=unzip_dir,
-        lim_perc_cloud=50,
+        lim_perc_cloud=cloud_min,
         level="LEVEL2A",
+        bands=bands,
         dry_run=False)
     
-    assert len(zip_files) == 1
+    assert len(downloaded) == 1
 
     # # simulate old version
     # for f in unzip_files:
@@ -73,20 +64,22 @@ def test_download(output_dir):
     #         f.move(new_file)
 
     # it should download duplicates (again)
-    zip_files, unzip_files = maja_download(
+    downloaded, unzip_files = maja_download(
         tile=tile,
         start_date=start_date,
         end_date=end_date,
         zip_dir=zip_dir,
         unzip_dir=unzip_dir,
-        lim_perc_cloud=60,
+        lim_perc_cloud=cloud_max,
         level="LEVEL2A",
+        bands=bands,
         dry_run=False)
     
-    assert len(zip_files) == 2
+    assert len(downloaded) == 2
+    assert len(unzip_files) == 2
     assert sum([f.exists() for f in unzip_files])==1
     # check if one of the files is considered as merged
-    df = get_local_maja_files(zip_dir=zip_dir, unzip_dir=unzip_dir)
+    df = get_local_maja_files(unzip_dir)
     assert len(df.merged_id.drop_duplicates()) == 1
 
     # nothing should be downloaded
@@ -96,8 +89,9 @@ def test_download(output_dir):
         end_date=end_date,
         zip_dir=zip_dir,
         unzip_dir=unzip_dir,
-        lim_perc_cloud=60,
+        lim_perc_cloud=cloud_max,
         level="LEVEL2A",
+        bands=bands,
         dry_run=False)
     
     assert len(zip_files) == 0
@@ -111,8 +105,9 @@ def test_download(output_dir):
         end_date=end_date,
         zip_dir=zip_dir,
         unzip_dir=unzip_dir,
-        lim_perc_cloud=50,
+        lim_perc_cloud=cloud_min,
         level="LEVEL2A",
+        bands=bands,
         dry_run=False)
     
     assert len(zip_files) == 0